# Conflicts:
#	ChaosandDeduction/UserSettings/EditorUserSettings.asset

+++ ChaosandDeduction/Assets/Scripts/Character/PlayerManager.cs
@@ -7,7 +7,7 @@ using Mirror;
 public struct PlayerData
 {
     public Alignment alignment;
-    public int modelIndex;
+    public byte modelIndex;
 }
 
 //  Class Attributes ----------------------------------
